more test framework modifications
note:
1. webpack might be the wrong approach.
2. consider using 'lab' instead.
